Converted milliseconds timestamps to 64 bit

+++ src/modules/datetime.c
@@ -8,19 +8,19 @@
 #include "modules/clock.h"
 #include "modules/config.h"
 
-uint32_t milliseconds = 0;
+uint64_t milliseconds = 0;
 
 void datetime_init(void)
 {
 	//TODO: Load timestamp from flash :D
 }
 
-uint32_t datetime_get_milliseconds()
+uint64_t datetime_get_milliseconds()
 {
 	return milliseconds;
 }
 
-void datetime_set_time(uint32_t new_milliseconds)
+void datetime_set_time(uint64_t new_milliseconds)
 {
 	milliseconds = new_milliseconds;
 }
